# perf(authentication-local): migrate to higher performance bcrypt implement