--- a/packages/authentication-local/package.json
+++ b/packages/authentication-local/package.json
@@ -55,12 +55,11 @@
     "@feathersjs/authentication": "^4.5.8",
     "@feathersjs/errors": "^4.5.8",
     "@feathersjs/feathers": "^4.5.8",
-    "bcryptjs": "^2.4.3",
+    "@node-rs/bcrypt": "^0.5.0",
     "debug": "^4.1.1",
     "lodash": "^4.17.19"
   },
   "devDependencies": {
-    "@types/bcryptjs": "^2.4.2",
     "@types/debug": "^4.1.5",
     "@types/lodash": "^4.14.158",
     "@types/mocha": "^8.0.0",