[@node-rs/bcrypt](https://www.npmjs.com/package/@node-rs/bcrypt) is `4~8x` faster than `bcryptjs`.

And compare to [node bcrypt](https://www.npmjs.com/package/bcrypt), `@node-rs/bcrypt` has no `postinstall` download scripts nor `node-gyp`.


Benchmark:

- [@node-rs/bcrypt](https://www.npmjs.com/package/@node-rs/bcrypt)
- [node bcrypt](https://www.npmjs.com/package/bcrypt)
- [bcryptjs](https://www.npmjs.com/package/bcryptjs)

```
@node-rs/bcrypt x 70.38 ops/sec ±1.40% (32 runs sampled)
node bcrypt x 63.88 ops/sec ±0.95% (30 runs sampled)
bcryptjs x 12.23 ops/sec ±4.90% (10 runs sampled)
Async hash round 10 bench suite: Fastest is @node-rs/bcrypt
@node-rs/bcrypt x 16.55 ops/sec ±2.33% (11 runs sampled)
node bcrypt x 14.62 ops/sec ±1.48% (11 runs sampled)
bcryptjs x 3.12 ops/sec ±1.60% (6 runs sampled)
Async hash round 12 bench suite: Fastest is @node-rs/bcrypt
@node-rs/bcrypt x 3.65 ops/sec ±5.08% (6 runs sampled)
node bcrypt x 3.50 ops/sec ±2.14% (6 runs sampled)
bcryptjs x 0.84 ops/sec ±12.31% (5 runs sampled)
Async hash round 14 bench suite: Fastest is @node-rs/bcrypt
@node-rs/bcrypt x 18.78 ops/sec ±0.86% (12 runs sampled)
node bcrypt x 16.28 ops/sec ±3.55% (11 runs sampled)
bcryptjs x 3.70 ops/sec ±2.64% (6 runs sampled)
Async verify bench suite: Fastest is @node-rs/bcrypt
```